ffmpeg 3.0 / 3.3 support by using build tags
fixes #38
